hash zala kadhacha

+++ userdata/views.py
@@ -2,17 +2,27 @@ from django.shortcuts import render, redirect
 from .models import *
 from .forms import *
 from django.contrib import messages
+from django.contrib.auth.decorators import login_required
 
+@login_required
+def dummy(request):
+    return redirect(home, request.user.id)
 
-def home(request):
+@login_required
+def home(request, pk):
+    # notes = note.objects.all()
+    user = User.objects.get(id=pk)
+    if user.id == request.user.id:
+        noten = user.note_set.all()
+        context = {
+            'title': 'Home',
+            'notes': noten
+        }
+        return render(request, 'userdata/home.html', context)
+    else:
+        return redirect(home, request.user.id)
 
-    context = {
-        'title': 'Home',
-        'notes': note.objects.all()
-    }
-    return render(request, 'userdata/home.html', context)
-
-
+@login_required
 def create_note(request):
     form = noteForm(request.POST or None, request.FILES or None)
     # user = note.author.id()
@@ -22,7 +32,7 @@ def create_note(request):
         obj.save()
         form = noteForm()
         messages.success(request, 'Your note has been added successfully!')
-        return redirect(home)
+        return redirect(home, request.user.id)
         
 
     context = {
@@ -31,22 +41,24 @@ def create_note(request):
     }
     return render(request, 'userdata/createnote.html', context)
 
-
+@login_required
 def view_note(request, pk):
     notes = note.objects.get(id=pk)
+    if notes.author.id == request.user.id :
+        if request.method == 'POST':
+            notes.delete()
+            messages.warning(request, 'Your note has been Deleted successfully!')
+            return redirect(home, request.user.id)
 
-    if request.method == 'POST':
-        notes.delete()
-        messages.warning(request, 'Your note has been Deleted successfully!')
-        return redirect(home)
+        context = {
+            'title': 'View Note',
+            'notes': notes,
+        }
+        return render(request, 'userdata/viewnote.html', context)
+    else:
+        return redirect(home, request.user.id)
 
-    context = {
-        'title': 'View Note',
-        'notes': notes,
-    }
-    return render(request, 'userdata/viewnote.html', context)
-
-
+@login_required
 def update_note(request, pk):
     notes = note.objects.get(id=pk)
     form = noteForm(instance=notes)
@@ -55,7 +67,7 @@ def update_note(request, pk):
         if form.is_valid():
             form.save()
             messages.success(request, 'Your note has been updated successfully!')
-            return redirect('/')
+            return redirect('home', request.user.id)
 
     context = {
         'title': 'Edit Note',
